Fix Gavin's action responses in BGEE.
In BGEE, SELECT_ACTION4-7 does not play on move as in BG2, but on
multiple clicks. Most NPC mods however assume that, because these mods
were written based on TuTu/BGT/EET which are all BG1-in-BG2 setups and
BGEE's SNDSLOT.IDS does not acknowledge the issue either.

2.6 patch named 4 priorly-unused slots in the cre file as
BGEE_ACTION4-7 for these actions. This patch moves Gavin's appropriate
lines to those slots.